pyproject.toml
[tool.poetry]
name = "lextract"
version = "0.0.1"
description = ""
authors = ["Frankie Robertson <frankie@robertson.name>"]
license = "Apache-2.0"
include = ["lextract/py.typed"]

[tool.poetry.dependencies]
python = "^3.7"
click = ">=6.7"
plumbum = "^1.6.6"
pyahocorasick = { git = "https://github.com/frankier/pyahocorasick.git", branch = "stiff-2018-09-20-3", optional = true }
opencc = "^0.2"
finntk = ">=0.0.72"
conllu = "^1.1"
click_log = "^0.3.2"
sqlalchemy_utils = "^0.36.5"
wikiparse = { git = "https://github.com/frankier/wikiparse.git", rev = "da7766fb112109873184bfd29d700a33b1385c19" }
numpy = "^1.19.2"
scikit-learn = "^0.23.2"
bitarray = "^1.5.3"

[tool.poetry.extras]
auto = ["pyahocorasick"]

[tool.poetry.dev-dependencies]
flake8 = "^3.5.0"
pre-commit = "^1.10.2"
pytest = "^4.1.1"
mypy = "0.770"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
